Revert common-lib-locator changes & deprecate
