--- library/common-lib-locator/api/common-lib-locator.klib.api
@@ -1,16 +1,0 @@
-// Klib ABI Dump
-// Targets: [androidNativeArm32, androidNativeArm64, androidNativeX64, androidNativeX86]
-// Rendering settings:
-// - Signature version: 2
-// - Show manifest properties: true
-// - Show declarations: true
-
-// Library unique name: <io.matthewnelson.kmp-tor:common-lib-locator>
-final class io.matthewnelson.kmp.tor.common.lib.locator/KmpTorLibLocator { // io.matthewnelson.kmp.tor.common.lib.locator/KmpTorLibLocator|null[0]
-    final object Companion { // io.matthewnelson.kmp.tor.common.lib.locator/KmpTorLibLocator.Companion|null[0]
-        final fun errorMsg(): kotlin/String // io.matthewnelson.kmp.tor.common.lib.locator/KmpTorLibLocator.Companion.errorMsg|errorMsg(){}[0]
-        final fun find(kotlin/String): io.matthewnelson.kmp.file/File? // io.matthewnelson.kmp.tor.common.lib.locator/KmpTorLibLocator.Companion.find|find(kotlin.String){}[0]
-        final fun isInitialized(): kotlin/Boolean // io.matthewnelson.kmp.tor.common.lib.locator/KmpTorLibLocator.Companion.isInitialized|isInitialized(){}[0]
-        final fun require(kotlin/String): io.matthewnelson.kmp.file/File // io.matthewnelson.kmp.tor.common.lib.locator/KmpTorLibLocator.Companion.require|require(kotlin.String){}[0]
-    }
-}
