+++ library/common-lib-locator/src/androidInstrumentedTest/kotlin/io/matthewnelson/kmp/tor/common/lib/locator/KmpTorLibLocatorAndroidTest.kt
@@ -1,0 +1,28 @@
+/*
+ * Copyright (c) 2025 Matthew Nelson
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     https://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ **/
+package io.matthewnelson.kmp.tor.common.lib.locator
+
+import kotlin.test.Test
+import kotlin.test.assertTrue
+
+class KmpTorLibLocatorAndroidTest {
+
+    @Test
+    fun givenAndroidRuntime_whenCheckInitialized_thenIsTrue() {
+        @Suppress("DEPRECATION")
+        assertTrue(KmpTorLibLocator.isInitialized())
+    }
+}
