--- test-android/api/test-android.klib.api
@@ -1,8 +1,0 @@
-// Klib ABI Dump
-// Targets: [androidNativeArm32, androidNativeArm64, androidNativeX64, androidNativeX86]
-// Rendering settings:
-// - Signature version: 2
-// - Show manifest properties: true
-// - Show declarations: true
-
-// Library unique name: <io.matthewnelson.kmp-tor:test-android>
