+++ test-android/build.gradle.kts
@@ -19,28 +19,12 @@ plugins {
 
 repositories { google() }
 
-private val jniLibs = project.projectDir
-    .resolve("src")
-    .resolve("androidMain")
-    .resolve("jniLibs")
-
 kmpConfiguration {
     configure {
         androidLibrary(namespace = "io.matthewnelson.kmp.tor.common.test") {
             android {
                 defaultConfig {
                     testInstrumentationRunner = "androidx.test.runner.AndroidJUnitRunner"
-
-                    packaging {
-                        jniLibs.useLegacyPackaging = true
-                    }
-
-                    ndk {
-                        abiFilters.add("arm64-v8a")
-                        abiFilters.add("armeabi-v7a")
-                        abiFilters.add("x86")
-                        abiFilters.add("x86_64")
-                    }
                 }
             }
 
@@ -57,61 +41,12 @@ kmpConfiguration {
             }
         }
 
-        androidNativeAll()
-
         common {
-            sourceSetMain {
-                dependencies {
-                    implementation(project(":library:common-lib-locator"))
-                }
-            }
             sourceSetTest {
                 dependencies {
                     implementation(kotlin("test"))
                 }
             }
         }
-
-        kotlin {
-            val buildDir = project.layout
-                .buildDirectory
-                .get()
-                .asFile
-
-            val nativeTestBinaryTasks = listOf(
-                "Arm32" to "armeabi-v7a",
-                "Arm64" to "arm64-v8a",
-                "X64" to "x86_64",
-                "X86" to "x86",
-            ).mapNotNull { (arch, abi) ->
-                val nativeTestBinariesTask = project
-                    .tasks
-                    .findByName("androidNative${arch}TestBinaries")
-                    ?: return@mapNotNull null
-
-                val abiDir = jniLibs.resolve(abi)
-                if (!abiDir.exists() && !abiDir.mkdirs()) throw RuntimeException("mkdirs[$abiDir]")
-
-                val testExecutable = buildDir
-                    .resolve("bin")
-                    .resolve("androidNative$arch")
-                    .resolve("debugTest")
-                    .resolve("test.kexe")
-
-                nativeTestBinariesTask.doLast {
-                    testExecutable.copyTo(abiDir.resolve("libTestExec.so"), overwrite = true)
-                }
-
-                nativeTestBinariesTask
-            }
-
-            project.tasks.all {
-                if (!name.startsWith("merge")) return@all
-                if (!name.endsWith("JniLibFolders")) return@all
-                nativeTestBinaryTasks.forEach { t -> dependsOn(t) }
-            }
-        }
     }
 }
-
-tasks.findByName("clean")?.apply { jniLibs.deleteRecursively() }
