--- test-android/src/androidNativeTest/kotlin/io/matthewnelson/kmp/tor/common/test/KmpTorLibLocatorNativeTest.kt
@@ -1,27 +1,0 @@
-/*
- * Copyright (c) 2025 Matthew Nelson
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- **/
-package io.matthewnelson.kmp.tor.common.test
-
-import kotlin.test.Test
-
-/**
- * Test executable is compiled, packaged into jniLibs, then run via androidInstrumentedTest
- * */
-class KmpTorLibLocatorNativeTest: KmpTorLibLocatorAndroidRuntimeTest() {
-
-    @Test
-    fun stub() {}
-}
