--- test-android/src/androidUnitTest/kotlin/io/matthewnelson/kmp/tor/common/test/KmpTorLibLocatorUnitTest.kt
@@ -1,28 +1,0 @@
-/*
- * Copyright (c) 2024 Matthew Nelson
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- **/
-package io.matthewnelson.kmp.tor.common.test
-
-import io.matthewnelson.kmp.tor.common.lib.locator.KmpTorLibLocator
-import kotlin.test.Test
-import kotlin.test.assertFalse
-
-class KmpTorLibLocatorUnitTest {
-
-    @Test
-    fun givenAndroidUnitTest_whenIsInitialized_thenIsFalse() {
-        assertFalse(KmpTorLibLocator.isInitialized())
-    }
-}
