--- test-android/src/commonTest/kotlin/io/matthewnelson/kmp/tor/common/test/KmpTorLibLocatorAndroidRuntimeTest.kt
@@ -1,37 +1,0 @@
-/*
- * Copyright (c) 2025 Matthew Nelson
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     https://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- **/
-package io.matthewnelson.kmp.tor.common.test
-
-import io.matthewnelson.kmp.file.File
-import io.matthewnelson.kmp.tor.common.lib.locator.KmpTorLibLocator
-import kotlin.test.Test
-import kotlin.test.assertNotNull
-import kotlin.test.assertTrue
-
-abstract class KmpTorLibLocatorAndroidRuntimeTest {
-
-    @Test
-    fun givenAndroidRuntime_whenIsInitialized_thenIsTrue() {
-        assertTrue(KmpTorLibLocator.isInitialized())
-    }
-
-    @Test
-    fun givenAndroidRuntime_whenFindLibTestExec_thenIsNotNull() {
-        assertNotNull(findLibTestExec())
-    }
-
-    protected fun findLibTestExec(): File? = KmpTorLibLocator.find("libTestExec.so")
-}
